update bundle scripts, firebot and smokebot to use directory paths without _64 (#1011)
* firebot: remove _64

* smokebot: remove _64

* cfastbot: remove size from smokeview builds, use 64 for cfast builds

* bundle scripts: remove 64 from bundle scripts

* cfast bundle: remove 64 from cfast bundle scripts

* scripts: remove _64 from web build scdripts

* scripts: correction to bundle scripts ( test_ -> _test)

* bundle scripts: add option to build smokeview nightly bundles using a specified hash and revision

* bundle script: add variable to control smokeview bundle upload

* bundle scripts: corrections to smv nightly bundle scripts

* bundle scripts: comment some lines in smv nightly bundle batch file

* bundle scripts: add option to nightly bundle fds-smv bundle scripts to build with specified revision and hash

* bundle scripts: corrections to windows fds/smv bundle script

* bundle scripts: move getopt.bat from fds to bot repo

* bundle scripts: add upload parameter to nightly smv bundle script

* bundle scripts: minor edits - add error message if bundle is not created

* bundle scripts: add parameters to fds/smv nightly bundle script for specifying revisions and hashes

* bundle scripts: remove _64

* scripts: remove _64

* scripts: remove _64

* scripts: remove unused files

* bundle scripts: checkout a specified branch first if bundle branch variable exists (fds and/or smv repos)

* bundle: remove smv directory

* smokebot: rename stage output files

* cfast bundle: remove _64 from cfast bundle scripts

* smokebot: minor formatting edits

* smokebot: minor formatting edits

* scripts: remove unused script - check_fds_images.sh

* firebot: fix some typos - correct some stage names

* smokebot: minor comment edits

---------

+++ Bundlebot/nightly/BuildSmvNightly.sh
@@ -2,6 +2,14 @@
 curdir=`pwd`
 BUILDING_nightly=$1
 
+UPLOAD=1
+S_HASH=
+S_REVISION=
+
+#UPLOAD=
+#S_HASH=2f257722a
+#S_REVISION=SMV-6.10.5-249
+
 #*** determine platform script is running on
 
 platform=linux
@@ -43,7 +51,7 @@ if [ "$BUILDING_release" == "" ]; then
   cd $reporoot/bot/Bundlebot/nightly/output
   outdir=`pwd`
   cd $reporoot/bot/Bundlebot/nightly
-  ./get_hash_revisions.sh $outdir >& $outdir/stage1_hash
+  ./get_hash_revisions.sh $outdir $S_HASH $S_REVISION >& $outdir/stage1_hash
   smv_hash=`head -1 $outdir/SMV_HASH`
 else
   cd $reporoot/bot/Bundlebot/release/output
@@ -73,7 +81,7 @@ echo "*** smv_revision: $smv_revision"
 echo "*** building libraries"
 
 # build libraries
-cd $reporoot/smv/Build/LIBS/${comp}_${platform}_64
+cd $reporoot/smv/Build/LIBS/${comp}_${platform}
 ./make_LIBS.sh >& $outdir/stage3_LIBS
 
 echo "*** building applications"
@@ -81,31 +89,40 @@ echo "*** building applications"
 progs="background flush pnginfo smokediff fds2fed smokezip wind2fds"
 
 for prog in $progs; do 
-  if [ -d $reporoot/smv/Build/$prog/${comp}_${platform}_64 ]; then
-    cd $reporoot/smv/Build/$prog/${comp}_${platform}_64
+  if [ -d $reporoot/smv/Build/$prog/${comp}_${platform} ]; then
+    cd $reporoot/smv/Build/$prog/${comp}_${platform}
     echo "*** building $prog"
     ./make_${prog}.sh >& $outdir/stage4_$prog
   fi
 done
 
 echo "*** building smokeview"
-cd $reporoot/smv/Build/smokeview/${comp}_${platform}_64
+cd $reporoot/smv/Build/smokeview/${comp}_${platform}
 ./make_smokeview.sh $BUILDTYPE >& $outdir/stage5_smokeview
 
 echo "*** bundling smokeview"
 
 $reporoot/bot/Bundlebot/nightly/assemble_smvbundle.sh $BUILDTYPE2 $smv_revision $basereporoot >& $outdir/stage6_bundle
 
-
-echo "*** uploading smokeview bundle"
-
-FILELIST=`gh release view SMOKEVIEW_TEST  -R github.com/$GHOWNER/test_bundles | grep SMV | grep -v FDS | grep $platform2 | awk '{print $2}'`
-for file in $FILELIST ; do
-  gh release delete-asset SMOKEVIEW_TEST $file -R github.com/$GHOWNER/test_bundles -y
-done
-
 uploaddir=.bundle/uploads
-$reporoot/bot/Bundlebot/nightly/upload_smvbundle.sh $uploaddir ${smv_revision}_${platform2}.sh     $basereporoot/bot/Bundlebot/nightly --clobber
+if [ -e $uploaddir/${smv_revision}_${platform2}.sh ]; then
+  echo smv bundle: $uploaddir/${smv_revision}_${platform2}.sh
+else
+  echo ***error: smv bundle: $uploaddir/${smv_revision}_${platform2}.sh failed to be created
+fi
 
-echo "*** upload complete"
+
+
+if [ "$UPLOAD" != "" ]; then
+  echo "*** uploading smokeview bundle"
+
+  FILELIST=`gh release view SMOKEVIEW_TEST  -R github.com/$GHOWNER/test_bundles | grep SMV |   grep -v FDS | grep $platform2 | awk '{print $2}'`
+  for file in $FILELIST ; do
+    gh release delete-asset SMOKEVIEW_TEST $file -R github.com/$GHOWNER/test_bundles -y
+  done
+
+  $reporoot/bot/Bundlebot/nightly/upload_smvbundle.sh $uploaddir ${smv_revision}_${platform2}.sh     $basereporoot/bot/Bundlebot/nightly --clobber
+
+  echo "*** upload complete"
+fi
